Add files via upload
This is a data cleaning project with Python in Jupyter Notebook, and the datasets of this project were from Kaggle website (linked below). The data provide information on the housing prices in Bei Jing from years 2011 to 2016, and the features associated with the property. 

Since the data were fetched from a China's real estate website, they contain quite a few columns of  foreign characters. Hence, the main job is to remove all these foreign characters, as well as the unreasonable data. 

Link to data: https://www.kaggle.com/ruiqurm/lianjia?select=new.csv