Add files via upload
One stage pooled OLS regression was implemented to estimate the empirical relationship between the financial viability of public transit service and the socioeconomic and demographic variables across rural and urban areas in Canada. 
The rural dummy, with 0 represents urban region while 1 represents rural region, aimed to capture the difference in the demand and supply of transit between urban and rural areas in Canada.